WinForms WebView: Setting Height, Width, Size, Bounds invalid
When setting the properties for `Height`, `Width`, `Size`, or `Bounds` during initialization, the underlying values for `Height` or `Width` are set to 0, causing the control to not render as expected.

Values are set to 0 in the `void Control.UpdateBounds()` method, which requires a client rectangle; however, since the control has not yet been rendered, there is no `RECT` and the resulting values are (correctly) `{0, 0}`. The logic has been changed to avoid most of the overrides and instead handle the `Layout` event, which manually updates the bounds of the `WebViewControl` if it has been initialized. If the control has not yet been initialized, the values are stored in the `Height`, `Width`, `Size`, and `Bounds` properties of the control as expected, and the values are used during initialization to set the initial size of the child HWND. Updates to any of the `Height`, `Width`, `Size`, or `Bounds` properties trigger the `Layout` event and the value is synchronized to the child HWND.