Add support for FAVICON in samples
Frequent question regarding how to interact with WebView without scriptable object support. This example shows how to load elements out of the DOM, serialize them to JSON, deserialize them, and perform actions. In this case, the icon candidates are read from the DOM after it is ready, and the items are asynchronously loaded from the internet and converted into a resource to be used by the application.